Centralize duplicated code and detach event listeners
