+++ packages/connect-node/src/connect-http-transport.ts
@@ -55,12 +55,9 @@ import {
   readToEnd,
   write,
 } from "./private/io.js";
-import * as http from "http";
-import * as https from "https";
-import {
-  nodeHeaderToWebHeader,
-  webHeaderToNodeHeaders,
-} from "./private/web-header-to-node-headers.js";
+import type * as http from "http";
+import type * as https from "https";
+import { nodeHeaderToWebHeader } from "./private/web-header-to-node-headers.js";
 import { assert } from "./private/assert.js";
 import {
   compressedFlag,
@@ -73,6 +70,7 @@ import {
   connectCreateRequestHeaderWithCompression,
   connectValidateResponseWithCompression,
 } from "./connect-http2-transport.js";
+import { getNodeRequest, makeNodeRequest } from "./private/node-request.js";
 
 const messageFlag = 0b00000000;
 
@@ -122,17 +120,6 @@ export interface ConnectHttpTransportOptions {
   compressMinBytes?: number;
   readMaxBytes?: number;
   sendMaxBytes?: number;
-}
-
-interface NodeRequestOptions<
-  I extends Message<I> = AnyMessage,
-  O extends Message<O> = AnyMessage
-> extends Pick<ConnectHttpTransportOptions, "httpOptions"> {
-  // Unary Request
-  req: UnaryRequest<I>;
-
-  // Request body
-  payload: Uint8Array;
 }
 
 /**
@@ -409,59 +396,4 @@ export function createConnectHttpTransport(
       );
     },
   };
-}
-
-function nodeRequest(protocol: string) {
-  if (protocol.startsWith("http") || protocol.startsWith("https")) {
-    return protocol.includes("https") ? https.request : http.request;
-  }
-  throw new Error("Unsupported protocol");
-}
-
-function makeNodeRequest(options: NodeRequestOptions) {
-  return new Promise<http.IncomingMessage>((resolve, reject) => {
-    const endpoint = new URL(options.req.url);
-    const nodeRequestFn = nodeRequest(endpoint.protocol);
-    const request = nodeRequestFn(options.req.url, {
-      headers: webHeaderToNodeHeaders(options.req.header),
-      method: "POST",
-      path: endpoint.pathname,
-      signal: options.req.signal,
-      ...options.httpOptions,
-    });
-
-    request.on("error", (err) => {
-      reject(err);
-    });
-
-    request.on("response", (res) => {
-      return resolve(res);
-    });
-
-    request.write(options.payload);
-    request.end();
-  });
-}
-
-function getNodeRequest(
-  req: StreamingRequest,
-  httpOptions: http.RequestOptions | https.RequestOptions | undefined
-) {
-  return new Promise<http.ClientRequest>((resolve, reject) => {
-    const endpoint = new URL(req.url);
-    const nodeRequestFn = nodeRequest(endpoint.protocol);
-    const request = nodeRequestFn(req.url, {
-      headers: webHeaderToNodeHeaders(req.header),
-      method: "POST",
-      path: endpoint.pathname,
-      signal: req.signal,
-      ...httpOptions,
-    });
-    request.on("socket", (socket) => {
-      socket.on("connect", () => {
-        resolve(request);
-      });
-      socket.on("error", reject);
-    });
-  });
 }
